NewsPaper/requirements.txt
django
django-filter
django-allauth
python-dotenv
django-apscheduler
celery
redis
celery[redis]